refine storage state
