+++ app/page/info/activity.tsx
@@ -1,7 +1,7 @@
 import { useAccount, useNetwork, useTransactionLogs } from "~app/storage"
 import { getChainName } from "~packages/network/util"
 import address from "~packages/util/address"
-import { TransactionStatus, type TransactionLog } from "~storage/local"
+import { TransactionStatus, type TransactionLog } from "~storage/local/state"
 
 const explorerUrl = "https://jiffyscan.xyz/"
