+++ app/storage/messenger.ts
@@ -3,7 +3,7 @@ import type { Patch } from "immer"
 import { sendToBackground, type MessageName } from "@plasmohq/messaging"
 
 import { StorageAction } from "~background/messages/storage"
-import { type State } from "~storage/local"
+import type { State } from "~storage/local/state"
 
 export class StorageMessenger {
   public get(): Promise<State> {
